% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-c-mean.R
\name{calcBelongDenom__}
\alias{calcBelongDenom__}
\title{Intermediar step in the calculation of the belonging matric}
\usage{
calcBelongDenom__(centerid, alldistances, m)
}
\arguments{
\item{centerid}{An integer representing the column to use as reference in the distance matrix}

\item{alldistances}{A distance matrix}
}
\value{
m a float representing the fuzzyness degree
}
\description{
Intermediar step in the calculation of the belonging matric
}
\examples{
#no example provided, this is an internal function
}
