% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-c-mean.R
\name{calcBelongMatrix}
\alias{calcBelongMatrix}
\title{Calculate the belonging matrix according to a set of centroids, the observed data and the fuzzyness degree}
\usage{
calcBelongMatrix(centers, data, m)
}
\arguments{
\item{centers}{A matrix or a dataframe representing the centers of the clusters with p columns and k rows}

\item{data}{A dataframe or matrix representing the observed data with nrows and p columns}

\item{m}{An integer representing the fuzzyness degree}
}
\value{
A n X k matrix represening the belongings of each datapoint to each cluster
}
\description{
Calculate the belonging matrix according to a set of centroids, the observed data and the fuzzyness degree
}
\examples{
data <- matrix(c(1,4,2,5,3,6,7,8,9,10,11,12),nrow=6,ncol=2)
centers <- data[1:2,]
calcBelongMatrix(centers,data,1.5)
}
