% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-c-mean.R
\name{calcCentroids}
\alias{calcCentroids}
\title{Calculate the new centroids of the clusters based on the belonging matrix}
\usage{
calcCentroids(data, belongmatrix, m)
}
\arguments{
\item{data}{A dataframe or matrix representing the observed data with nrows and p columns}

\item{belongmatrix}{A n X k matrix giving for each observation n, its probability to belong to the cluster k}

\item{m}{An integer representing the fuzzyness degree}
}
\value{
A n X k matrix represening the belongings of each datapoint to each cluster
}
\description{
Calculate the new centroids of the clusters based on the belonging matrix
}
\examples{
data <- matrix(c(1,4,2,5,3,6,7,8,9,10,11,12),nrow=6,ncol=2)
centers <- data[1:2,]
belongmatrix <- calcBelongMatrix(centers,data,1.5)
newcenters <- calcCentroids(data,belongmatrix,1.5)
}
