% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-c-mean.R
\name{calcEuclideanDistance}
\alias{calcEuclideanDistance}
\title{Calculate the euclidean distance between a numeric matrix n X p and a numeric vector of length p}
\usage{
calcEuclideanDistance(m, v)
}
\arguments{
\item{m}{A n X p matrix or dataframe with only numeric columns}

\item{v}{A numerci vector of length p}
}
\value{
A vector of length n giving the euclidean distance between all matrix row and the vector p
}
\description{
Calculate the euclidean distance between a numeric matrix n X p and a numeric vector of length p
}
\examples{
mat <- matrix(c(1,4,2,5,3,6),nrow=2,ncol=3)
v1 <- c(1,2,3)
calcEuclideanDistance(mat,v1)
}
