% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_help.R
\name{calcqualityIndexes}
\alias{calcqualityIndexes}
\title{calculate several clustering quality indexes (most of them come from fclust package)}
\usage{
calcqualityIndexes(data, belongmatrix, centers)
}
\arguments{
\item{data}{the original dataframe used fot the classification}

\item{belongmatrix}{The belonging matrix obtained at the end of the algorithm}

\item{centers}{The dataframe representing the centers of the clusters}
}
\description{
calculate several clustering quality indexes (most of them come from fclust package)
}
\examples{

}
