% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-c-mean.R
\name{calcSWCMBelongMatrix}
\alias{calcSWCMBelongMatrix}
\title{Calculate the belonging matrix (spatial version) according to a set of centroids, the observed data, the fuzzyness degree a neighbouring matrix and a spatial ponderation term}
\usage{
calcSWCMBelongMatrix(centers, data, neighbours, m, alpha)
}
\arguments{
\item{centers}{A matrix or a dataframe representing the centers of the clusters with p columns and k rows}

\item{data}{A dataframe or matrix representing the observed data with nrows and p columns}

\item{neighbours}{A list.w object describing the neighbours typically produced by the spdep package}

\item{m}{An integer representing the fuzzyness degree}

\item{alpha}{An integer representing the weight of the space in the analysis (0 is a typical fuzzy-c-mean algorithm)}
}
\value{
A n X k matrix represening the belongings of each datapoint to each cluster
}
\description{
Calculate the belonging matrix (spatial version) according to a set of centroids, the observed data, the fuzzyness degree a neighbouring matrix and a spatial ponderation term
}
\examples{
data <- matrix(c(1,4,2,5,3,6,7,8,9,10,11,12),nrow=6,ncol=2)
centers <- data[1:2,]
calcBelongMatrix(centers,data,1.5)
}
