% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-c-mean.R
\name{calcSWFCCentroids}
\alias{calcSWFCCentroids}
\title{Calculate the new centroids of the clusters based on the belonging matrix (spatial version)}
\usage{
calcSWFCCentroids(data, belongmatrix, neighbours, m, alpha)
}
\arguments{
\item{data}{A dataframe or matrix representing the observed data with nrows and p columns}

\item{belongmatrix}{A n X k matrix giving for each observation n, its probability to belong to the cluster k}

\item{neighbours}{A list.w object describing the neighbours typically produced by the spdep package}

\item{m}{An integer representing the fuzzyness degree}

\item{alpha}{An integer representing the weight of the space in the analysis (0 is a typical fuzzy-c-mean algorithm)}
}
\value{
A n X k matrix represening the belongings of each datapoint to each cluster
}
\description{
Calculate the new centroids of the clusters based on the belonging matrix (spatial version)
}
\examples{
data <- matrix(c(1,4,2,5,3,6,7,8,9,10,11,12),nrow=6,ncol=2)
centers <- data[1:2,]
belongmatrix <- calcBelongMatrix(centers,data,1.5)
newcenters <- calcCentroids(data,belongmatrix,1.5)
}
