% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_help.R
\name{mapClusters}
\alias{mapClusters}
\title{build some maps to visualize the results of the clustering}
\usage{
mapClusters(geodata, belongmatrix, undecided = NULL)
}
\arguments{
\item{belongmatrix}{The belonging matrix obtained at the end of the algorithm}
}
\description{
build some maps to visualize the results of the clustering
}
\examples{

}
