% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_help.R
\name{spatialdiag}
\alias{spatialdiag}
\title{calculate the moran I of each column of the belonging matrix and the multiple join count test of the most probablt cluster (spdep package)}
\usage{
spatialdiag(belongmatrix, nblistw, undecided = NULL)
}
\arguments{
\item{belongmatrix}{The belonging matrix obtained at the end of the algorithm}

\item{A}{list.w object describing the neighbours typically produced by the spdep package}
}
\description{
calculate the moran I of each column of the belonging matrix and the multiple join count test of the most probablt cluster (spdep package)
}
\examples{

}
