% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretation_help.R
\name{summarizeClusters}
\alias{summarizeClusters}
\title{Calculate some descriptive statistics of each group}
\usage{
summarizeClusters(data, belongmatrix, weighted = TRUE, dec = 3)
}
\arguments{
\item{data}{the original dataframe used fot the classification}

\item{belongmatrix}{The belonging matrix obtained at the end of the algorithm}
}
\description{
Calculate some descriptive statistics of each group
}
\examples{

}
